refactor: move the FAQ section under app/components
